+++ src/app/app.component.ts
@@ -1,23 +1,10 @@
-import { Component, OnInit } from '@angular/core';
-import { ProductsService } from './product/services/products.service';
-import { Book } from './product/model/book.model';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent implements OnInit {
+export class AppComponent {
   title = 'My Shop';
-  books: Book[];
-
-  constructor(private productService: ProductsService) {}
-
-  ngOnInit() {
-    this.getBooks();
-  }
-
-  private async getBooks() {
-    this.books = <Book[]>this.productService.getProducts();
-  }
 }