+++ src/app/cart/cart.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { SharedModule } from '../shared/shared.module';
+
+import { CartService } from './services/cart.service';
+
+import { CartItemComponent } from './components/cart-item/cart-item.component';
+import { CartListComponent } from './components/cart-list/cart-list.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    SharedModule
+  ],
+  declarations: [
+    CartListComponent,
+    CartItemComponent
+  ],
+  exports: [
+    CartListComponent
+  ],
+  providers: [
+    CartService
+  ]
+})
+export class CartModule { }