+++ src/app/product/model/product.model.ts
@@ -2,5 +2,6 @@ export interface Product {
     name: string;
     price: number;
     quantity?: number;
+    availableQuantity?: number;
     location?: string;
 }