+++ src/app/product/product.module.ts
@@ -1,17 +1,24 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { ProductsService } from './services/products.service';
+
 import { ProductComponent } from './components/product/product.component';
+import { ProductListComponent } from './components/product-list/product-list.component';
 
 @NgModule({
   imports: [
     CommonModule
   ],
   declarations: [
-    ProductComponent
+    ProductComponent,
+    ProductListComponent
   ],
   exports: [
-    ProductComponent
+    ProductListComponent
+  ],
+  providers: [
+    ProductsService
   ]
 })
 export class ProductModule { }